*: move extension implementation to RegisterBlocks
With the old implementation, the only way to get access to the Ext
methods for GPIO and CMU was the following:

    let periph = efm32gg11b820::Peripherals::take().unwrap();
    let mut cmu = periph.CMU;
    let gpio = periph.GPIO.split(cmu.constrain().split().gpio);

This can be problematic if it is desired to "steal" the peripherals
(e.g. in an interrupt handler) because the peripherals cannot be taken
multiple times. The recommended way to do this with libraries generated
by svd2rust is to use the `ptr` method as follows:

    let gpio = (unsafe { &*efm32gg11b820::GPIO::ptr() }).split(
        (unsafe { &*efm32gg11b820::CMU::ptr() })
            .constrain()
            .split()
            .gpio,
    );

This new implementation allows for both invocations by implementing Ext
on the underlying RegisterBlocks type instead of the top-level phantom
type.